# candidates-openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "api-companies",
        "description": "Microservices candidates",
        "version": "1.0.0",
        "x-logo": {
            "url": "https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png"
        }
    },
    "paths": {
...
        "/api/candidates/language": {
            "get": {
                "tags": [
                    "language"
                ],
                "summary": "Get Language",
                "operationId": "get_language_api_candidates_language_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "language"
                ],
                "summary": "Create Language",
                "operationId": "create_language_api_candidates_language_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Language"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidates/language/{id}": {
            "get": {
                "tags": [
                    "language"
                ],
                "summary": "Get Language",
                "operationId": "get_language_api_candidates_language__id__get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "language"
                ],
                "summary": "Put Language",
                "operationId": "put_language_api_candidates_language__id__put",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkExperienceUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidates/education": {
            "get": {
                "tags": [
                    "education"
                ],
                "summary": "Get Educations",
                "operationId": "get_educations_api_candidates_education_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "education"
                ],
                "summary": "Create Education",
                "operationId": "create_education_api_candidates_education_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Education"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidates/education/{id}": {
            "get": {
                "tags": [
                    "education"
                ],
                "summary": "Get Educations",
                "operationId": "get_educations_api_candidates_education__id__get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "education"
                ],
                "summary": "Put Educations",
                "operationId": "put_educations_api_candidates_education__id__put",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EducationUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/candidates/profile": {
            "get": {
                "tags": [
                    "profile"
                ],
                "summary": "Get Profiles",
                "operationId": "get_profiles_api_candidates_profile_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "profile"
                ],
                "summary": "Create Profile",
                "operationId": "create_profile_api_candidates_profile_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidates/profile/{id}": {
            "get": {
                "tags": [
                    "profile"
                ],
                "summary": "Get Profiles",
                "operationId": "get_profiles_api_candidates_profile__id__get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "profile"
                ],
                "summary": "Put Profiles",
                "operationId": "put_profiles_api_candidates_profile__id__put",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AcademicLevel": {
                "type": "string",
                "enum": [
                    "PREGRADO",
                    "ESPECIALIZACIÓN",
                    "MAESTRÍA",
                    "DOCTORADO",
                    "DIPLOMADO",
                    "CURSO",
                    "CERTIFICACIÓN"
                ],
                "title": "AcademicLevel"
            },
...
            "Education": {
                "properties": {
                    "start_date": {
                        "type": "string",
                        "title": "Start Date"
                    },
                    "end_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "End Date"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "educational_institution": {
                        "type": "string",
                        "title": "Educational Institution"
                    },
                    "degree": {
                        "$ref": "#/components/schemas/AcademicLevel"
                    },
                    "state": {
                        "$ref": "#/components/schemas/StateEducation"
                    },
                    "candidate_id": {
                        "type": "string",
                        "title": "Candidate Id"
                    }
                },
                "type": "object",
                "required": [
                    "start_date",
                    "title",
                    "educational_institution",
                    "degree",
                    "state",
                    "candidate_id"
                ],
                "title": "Education"
            },
            "EducationUpdate": {
                "properties": {
                    "start_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Start Date"
                    },
                    "end_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "End Date"
                    },
                    "title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Title"
                    },
                    "educational_institution": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Educational Institution"
                    },
                    "degree": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Degree"
                    },
                    "state": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "State"
                    },
                    "candidate_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Candidate Id"
                    }
                },
                "type": "object",
                "title": "EducationUpdate"
            },
            "EmploymentType": {
                "type": "string",
                "enum": [
                    "JORNADA COMPLETA",
                    "JORNADA PARCIAL",
                    "CONTRATO TEMPORAL",
                    "PROFESIONAL INDEPENDIENTE",
                    "CONTRATO DE PRÁCTICAS"
                ],
                "title": "EmploymentType"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Language": {
                "properties": {},
                "type": "object",
                "title": "Language"
            },
            "Profile": {
                "properties": {},
                "type": "object",
                "title": "Profile"
            },
            "ResponseDto": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "items": {},
                                "type": "array"
                            },
                            {
                                "type": "object"
                            },
                            {
                                "type": "string"
                            }
                        ],
                        "title": "Data"
                    },
                    "errors": {
                        "type": "string",
                        "title": "Errors"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "data",
                    "errors"
                ],
                "title": "ResponseDto"
            },
...
            "StateEducation": {
                "type": "string",
                "enum": [
                    "EN CURSO",
                    "FINALIZADO",
                    "CANCELADO",
                    "APLAZADO",
                    "PENDIENTE GRADO"
                ],
                "title": "StateEducation"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
            "WorkExperienceUpdate": {
                "properties": {
                    "company": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Company"
                    },
                    "job_title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Job Title"
                    },
                    "start_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Start Date"
                    },
                    "end_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "End Date"
                    },
                    "functions": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Functions"
                    },
                    "job_type": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/EmploymentType"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "is_actual": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Actual"
                    },
                    "candidate_id": {
                        "type": "integer",
                        "title": "Candidate Id"
                    }
                },
                "type": "object",
                "required": [
                    "candidate_id"
                ],
                "title": "WorkExperienceUpdate"
            }
        }
    }
}